requirements.txt
pywaggle
pyserial

